+++ RELEASE_NOTES.md
@@ -1,6 +1,8 @@
 # Release notes
 
-## New in git-machete 3.34.1
+## New in git-machete 3.35.0
+
+- added: `git machete traverse --sync-github-prs`/`--sync-gitlab-mrs` suggests creating PRs/MRs for untracked branches (suggested by @bradneuman and @tir38)
 
 - fixed: `yq` answer to retargeting a PR/MR in `traverse` was treated just as `y`/`yes` (no quitting)